RESULTS UPDATED

+++ src/main/java/ub/fet/smartschool/repository/ResultRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ub.fet.smartschool.model.Result;
 
 @Repository
-public interface ResultRepository extends JpaRepository<Result,Long> {
+public interface ResultRepository extends JpaRepository<Result,Integer> {
 }
